refactored ego4d -> ego4d_forecasting
