+++ ego4d_forecasting/models/stem_helper.py
